[feat/#10] Implement WebViewScreen
