+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.MisikAndroid"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.nexters.misik.webview.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MisikAndroid">
